Move graph module from lib or vendor directory to app directory.
Because not autoloading lib directory at development mode.
